* develop: (21 commits)
  Removed redundant Event class. Fixed bug in BaseRecyclerViewFragment.java.
  Added "more button" support.
  Added listapps BRICK layout support.
  Fixed bug in activity_main.xml.
  added generic dialogs and some java docs corrected a bug where the cache wasn't work properly
  Finished MainActivityFragment.java.
  Turned WIdgetArgs dynamic.
  changed the style to match the app style fixed the bug where progress dialog box didn't dismiss if login failed
  leak solved
  Update user settings on ws
  added progress dialog box
  getUserInfo implemented
  Relog if Refresh token is not valid
  The correct errors are now displayed
  corrected some login errors corrected proGuard rules
  My account added to navigation drawer
  added my account activity
  Implemented the method to refresh the access token
  Added OAuth.java and CheckUserCredentialsRequest.java web services.
  dbz
  ...

# Conflicts:
#	model/src/main/java/cm/aptoide/pt/model/v7/Type.java
#	v8engine/src/main/java/cm/aptoide/pt/v8engine/activity/AptoideFragmentActivity.java
#	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/displayable/Displayable.java

+++ model/src/main/java/cm/aptoide/pt/model/v7/listapp/App.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2016.
- * Modified by Neurophobic Animal on 27/04/2016.
+ * Modified by Neurophobic Animal on 10/05/2016.
  */
 
 package cm.aptoide.pt.model.v7.listapp;
@@ -43,7 +43,7 @@ public class App {
 		@Data
 		public static class Rating {
 
-			private int avg;
+			private float avg;
 			private int total;
 			private List<Vote> votes;
